Consider line height to calculate page breaks + ignore br on top of page
If an element has a very long text, page break calculation takes that
into account. Note: we still don't consider margins on the calculation.

Also done on this commit: ignore any empty line on top of pages (we need
to do this due to a limitation on :after/:before pseudo elements on top
of <br> (empty lines on Etherpad)