Do not break pages in the middle of sequence of transitions
When there is a sequence of transitions, the last 2 of them belong to a
block; the others are considered individual lines and can be split into
separated pages.

This breaks the spec when we only have transitions on the script. Due to
a bug on Final Draft, when pagination places a transition on top of the
page, they pull the last element of previous page for the next page (to
avoid having a transition on top of the page), but they forgot to verify
if last element of previous page is also a transition (which would make
the element-pulling useless). We'll handle this pagination rule later.