Refactor logic to paginate script: look previous element instead of next
Only check if there is a block when an element is identified as "to be
moved to next page". This is more efficient AND is more aligned to the
logic on Final Draft, so we need to implement less rules in order to
achieve the same results of FD pagination.