Use correct height for text after page break when element is split
When element is split, the resulting two-halves might increase the total
number of lines of the element. For example, if it has two sentences
with 75 chars each, the two sentences together need only 3 lines to fit,
while if they are split it would need 4 lines to fit them (2 for each
sentence). This commit fixes this scenario.