Display MORE/CONT'D when page break is between elements
To be able to do that, we had to change the way page break between
elements are displayed: instead of simply adding a class to the <div>,
we now add a line attribute and build DOM when reading the line. This
made the code more uniform (split and non-split page breaks use
attributes), but added some complexity.