Transform splitPageBreak from char attrib to line attrib
This fixes the bug where caret was being moved to previous line when
placed at the first position of a page.

In order to do this change we had to insert a "\n" at the position where
element is being split between pages. We still need to fix some issues
related to this (HUGE) change, so this commit is still WIP.