Remove 2 chars when user presses DEL or BACKSPACE on edge of split line
If user presses BACKSPACE and caret is on beginning of 2nd half of split
line, it should merge splits + remove last char of 1st half; if caret is
at the end of 1st half and user presses DELETE, it should merge splits +
remove 1st char of 2nd half.

In order to do that, and to manage dependencies with ep_script_elements,
it was necessary to change ep.json of both plugins -- have a separated
part for aceKeyEvent and aceDomLineProcessLineAttributes.