Fix page numbering
If first page break of cycle was repaginated, it would be replaced when
pagination cleanup was run, so when getting next page number the
reference to the original line would be broken.

Also fix issue of a page break left uncleaned when its line was between
original line number (before pagination) and final page number (after
pagination).