[refactor] Remove unused CSS for larger screens
On TK, we only show the editor on the max-width mode (no side borders),
which is `642px`. So it doesn't make sense to keep any CSS style for
screens larger than that.

Let's move this code back when we implement the feature to show editor
on other page views (full-page, for example).