[feat] Improve speed to calculate page breaks
Don't clone lines on each pagination cycle; instead, use their line
sized (provided by ep_script_line_size) to calculate page breaks on the
fly. This allows us to remove a lot of code related to "cloned lines",
"helper lines", etc.

Also:

- optimize repagination: stop paginating when pagination cycle reaches a
point where the existing page break is on the same spot of the
calculated page break;
- use Set (instead of an object) to store lines changed on
`paginationLinesChanged`
- create helper methods for tests:
  - waitToHaveAnyPageBreak
  - waitToHaveAnySplitPageBreak
  - waitToHaveAnyNonSplitPageBreak
  - waitToHaveNNonSplitPageBreaks
  - waitToHaveNSplitPageBreaks
- un-pend tests and fix them so they can validate pagination again:
  - [x] basicPagination.js: adjusted and all tests passing
  - [x] calculatingPageNumber.js: adjusted and all tests passing
  - [ ] elementBlocks.js: not adjusted
  - [x] enablePagination.js: adjusted and all tests passing
  - [-] lineNumberHeight.js (should we remove this suite? We don't allow
        users to show line numbers anymore...)
  - [ ] otherMoreContd.js: not adjusted
  - [ ] repaginate.js: adjusted, but with tests failing
  - [ ] scroll.js: not adjusted
  - [ ] splitElements.js: adjusted, but with tests failing
- optimize tests: don't create a new pad for each test, use the same one
and clean pad content before starting each test. This can be even more
efficient if we use UNDO instead of cleaning the pad, but to do that we
need to change the tests in a lot of places, so this is not done on this
commit;
- mark failing tests with `// FIXME`;

Implement part of https://trello.com/c/dJp83kE0/1971.